# swagger-static/swagger.json
{"openapi":"3.0.0","paths":{"/user":{"post":{"operationId":"UserController_createUser","summary":"신규 사용자 추가","parameters":[],"requestBody":{"required":true,"description":"신규 사용자 정보","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"신규 사용자 추가","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"400":{"description":"잘못된 요청"}},"tags":["User API"]}},"/user/info":{"get":{"operationId":"UserController_getUserInfoByEmail","parameters":[{"name":"email","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["User API"]}},"/auth/github":{"get":{"operationId":"AuthController_gituhbLogin","parameters":[],"responses":{"200":{"description":""}}}},"/auth/github/callback":{"get":{"operationId":"AuthController_githubLoginCallback","parameters":[],"responses":{"200":{"description":""}}}},"/auth/refresh":{"post":{"operationId":"AuthController_refreshToken","parameters":[],"responses":{"201":{"description":""}}}},"/board":{"get":{"operationId":"BoardController_getPosts","summary":"전체 게시글 조회","parameters":[],"responses":{"200":{"description":"전체 게시글 조회","content":{"application/json":{"example":[{"post_id":1,"title":"게시물 제목","content":"게시물 내용","update_at":"2024-10-01T00:00:12","create_at":"2024-10-03T00:00:12","expired_at":"2024-10-06T00:00:12","view_cnt":15,"bookmarked_cnt":3,"is_open":true},{"post_id":2,"title":"게시물 제목2","content":"게시물 내용2","update_at":"2024-10-01T00:00:12","create_at":"2024-10-03T00:00:12","expired_at":"2024-10-06T00:00:12","view_cnt":13,"bookmarked_cnt":52,"is_open":true}]}}}}},"post":{"operationId":"BoardController_createPost","summary":"게시글 생성","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBoardDto"}}}},"responses":{"201":{"description":"게시글 생성","content":{"application/json":{"example":{"post_id":1,"title":"게시물 제목","content":"게시물 내용","update_at":"2024-10-01T00:00:12","create_at":"2024-10-03T00:00:12","expired_at":"2024-10-06T00:00:12","view_cnt":15,"bookmarked_cnt":3,"is_open":true}}}}}}},"/board/{id}":{"get":{"operationId":"BoardController_getPostById","summary":"게시글 상세 조회","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"게시글 상세 조회","content":{"application/json":{"example":{"post_id":1,"title":"게시물 제목","content":"게시물 내용","update_at":"2024-10-01T00:00:12","create_at":"2024-10-03T00:00:12","expired_at":"2024-10-06T00:00:12","view_cnt":15,"bookmarked_cnt":3,"is_open":true}}}}}},"patch":{"operationId":"BoardController_updatePost","summary":"게시글 수정","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateBoardDto"}}}},"responses":{"200":{"description":"게시글 수정","content":{"application/json":{"example":{"post_id":1,"title":"수정된 게시물 제목","content":"수정된 게시물 내용","update_at":"2024-10-01T00:00:12","create_at":"2024-10-03T00:00:12","expired_at":"2024-10-06T00:00:12","view_cnt":15,"bookmarked_cnt":3,"is_open":false}}}}}},"delete":{"operationId":"BoardController_deletePost","summary":"게시글 삭제","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"204":{"description":"게시글 삭제 완료"}}}},"/chat/rooms":{"post":{"operationId":"ChatController_createChatRoom","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateChatRoomDto"}}}},"responses":{"201":{"description":""}},"tags":["Chat API"]},"get":{"operationId":"ChatController_getChatRooms","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetChatRoomsDto"}}}},"responses":{"200":{"description":""}},"tags":["Chat API"]}},"/chat/rooms/{id}":{"get":{"operationId":"ChatController_getChatRoom","parameters":[],"responses":{"200":{"description":""}},"tags":["Chat API"]}},"/chat/lists/{id}":{"get":{"operationId":"ChatController_getChatList","parameters":[],"responses":{"200":{"description":""}},"tags":["Chat API"]}},"/public-data/post-categories":{"get":{"operationId":"PublicDataController_getPostCategories","summary":"전체 post category 데이터 조회","parameters":[],"responses":{"200":{"description":"전체 post category 데이터 조회","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PostCategoryResDto"}}}}},"404":{"description":"데이터가 없습니다."}},"tags":["Public Data API"]}},"/public-data/post-categories/{id}":{"get":{"operationId":"PublicDataController_getPostCategoryById","summary":"특정 post category 데이터 조회","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"특정 post category 데이터 조회","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostCategoryResDto"}}}}},"tags":["Public Data API"]}},"/public-data/position-categories":{"get":{"operationId":"PublicDataController_getPositionCategories","summary":"전체 position category 데이터 조회","parameters":[],"responses":{"200":{"description":"전체 position category 데이터 조회","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PositionCategoryResDto"}}}}}},"tags":["Public Data API"]}},"/public-data/position-categories/{id}":{"get":{"operationId":"PublicDataController_getPositionCategoryById","summary":"특정 position category 데이터 조회","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"특정 position category 데이터 조회","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PositionCategoryResDto"}}}},"404":{"description":"데이터가 없습니다."}},"tags":["Public Data API"]}},"/public-data/project-categories":{"get":{"operationId":"PublicDataController_getProjectCategories","summary":"전체 project category 데이터 조회","parameters":[],"responses":{"200":{"description":"전체 project category 데이터 조회","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProjectCategoryResDto"}}}}}},"tags":["Public Data API"]}},"/public-data/project-categories/{id}":{"get":{"operationId":"PublicDataController_getProjectCategoryById","summary":"특정 project category 데이터 조회","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"특정 project category 데이터 조회","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectCategoryResDto"}}}}},"tags":["Public Data API"]}},"/public-data/stack-categories":{"get":{"operationId":"PublicDataController_getStackCategories","summary":"전체 stack category 데이터 조회","parameters":[],"responses":{"200":{"description":"전체 stack category 데이터 조회","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/StackCategoryResDto"}}}}}},"tags":["Public Data API"]}},"/public-data/stack-categories/{id}":{"get":{"operationId":"PublicDataController_getStackCategoryById","summary":"특정 stack category 데이터 조회","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"특정 stack category 데이터 조회","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StackCategoryResDto"}}}}},"tags":["Public Data API"]}},"/public-data/career-categories":{"get":{"operationId":"PublicDataController_getCareerCategories","summary":"전체 career category 데이터 조회","parameters":[],"responses":{"200":{"description":"전체 career category 데이터 조회","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CareerCategoryResDto"}}}}}},"tags":["Public Data API"]}},"/public-data/career-categories/{id}":{"get":{"operationId":"PublicDataController_getCareerCategoryById","summary":"특정 career category 데이터 조회","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"특정 career category 데이터 조회","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CareerCategoryResDto"}}}}},"tags":["Public Data API"]}}},"info":{"title":"GrabbMe API Docs","description":"GrabbMe Backend API Documents","version":"1.0.0","contact":{}},"tags":[],"servers":[{"url":"api"}],"components":{"schemas":{"CreateUserDto":{"type":"object","properties":{"nickname":{"type":"string"},"email":{"type":"string"},"stack_category_id":{"type":"array","items":{"type":"number"}},"position_category_id":{"type":"number"},"project_category_id":{"type":"array","items":{"type":"number"}},"career_category_id":{"type":"number"}},"required":["nickname","email","stack_category_id","position_category_id","project_category_id","career_category_id"]},"User":{"type":"object","properties":{}},"CreateBoardDto":{"type":"object","properties":{}},"UpdateBoardDto":{"type":"object","properties":{}},"CreateChatRoomDto":{"type":"object","properties":{}},"GetChatRoomsDto":{"type":"object","properties":{}},"PostCategoryResDto":{"type":"object","properties":{"id":{"type":"number","example":1,"description":"post category id"},"post_category_name":{"type":"string","example":"프로젝트","description":"post category name"}},"required":["id","post_category_name"]},"PositionCategoryResDto":{"type":"object","properties":{"position_category_id":{"type":"number","examples":[1,2,3],"description":"position category id"},"name":{"type":"string","examples":["Frontend","Backend","Fullstack"],"description":"직군 이름"},"kor_name":{"type":"string","examples":["프론트엔드","백엔드","풀스택"],"description":"직군 한글 이름"},"abbreviation":{"type":"string","examples":["FE","BE","FS"],"description":"직군 약어"}},"required":["position_category_id","name","kor_name","abbreviation"]},"ProjectCategoryResDto":{"type":"object","properties":{"project_category_id":{"type":"number","examples":[1,2,3],"description":"project category id"},"name":{"type":"string","examples":["B2B","B2C","Fintech"],"description":"도메인 이름"},"kor_name":{"type":"string","examples":["B2B","B2C","핀테크"],"description":"도메인 한글 이름"},"description":{"type":"string","examples":["기업간","기업대 소비자","금융"],"description":"도메인 설명"}},"required":["project_category_id","name","kor_name","description"]},"StackCategoryResDto":{"type":"object","properties":{"stack_category_id":{"type":"number","examples":[1,2,3],"description":"stack category id"},"name":{"type":"string","examples":["React","Node.js","Django"],"description":"스택 이름"},"kor_name":{"type":"string","examples":["리액트","노드","장고"],"description":"스택 한글 이름"},"abbreviation":{"type":"string","examples":["FE","BE","BE"],"description":"스택 약어"}},"required":["stack_category_id","name","kor_name","abbreviation"]},"CareerCategoryResDto":{"type":"object","properties":{"career_category_id":{"type":"number","examples":[1,2,3],"description":"career category id"},"content":{"type":"string","example":"0년차","description":"career category name"}},"required":["career_category_id","content"]}}}}